pyproject.toml
[project]
name = "pieceofcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.3",
    "django-htmx>=1.21.0",
    "django-tailwind[reload]>=3.8.0",
    "django-stubs>=5.1.1",
    "django-unfold>=0.40.0",
]
